Rename duplicate modules
