+++ src/pe_source/pe_shodan.py
@@ -7,13 +7,12 @@ import threading
 import numpy
 
 from .data.pe_db.config import shodan_api_init
-from .data.pe_db.db_query import get_orgs
+from .data.pe_db.db_query_source import get_orgs
 from .data.shodan.shodan_search import run_shodan_thread
-
 
 class Shodan:
     """Fetch Shodan data."""
-
+    
     def __init__(self, orgs_list):
         """Initialize Shodan class."""
         self.orgs_list = orgs_list
@@ -24,7 +23,7 @@ class Shodan:
 
         # Get orgs from PE database
         pe_orgs = get_orgs()
-
+        
         # Filter orgs if specified
         if orgs_list == "all":
             pe_orgs_final = pe_orgs
@@ -38,10 +37,13 @@ class Shodan:
 
         # Get list of initialized API objects
         api_list = shodan_api_init()
+        chunked_orgs_list = numpy.array([])
 
-        # Split orgs into chunks. # of chunks = # of valid API keys = # of threads
-        chunk_size = len(api_list)
-        chunked_orgs_list = numpy.array_split(numpy.array(pe_orgs_final), chunk_size)
+        # Split orgs into chunks. # of chunks = # of valid API keys = # of threads 
+        # if list has any valid apis
+        if len(api_list) > 0:
+            chunk_size = len(api_list)
+            chunked_orgs_list = numpy.array_split(numpy.array(pe_orgs_final), chunk_size)
 
         i = 0
         thread_list = []
